Remove deprecated legacy HighThroughEngine
[sc-38557]
