+++ changelog.d/20250109_164900_kevin_remove_deprecated_htex.rst
@@ -1,0 +1,6 @@
+Removed
+^^^^^^^
+
+- Remove ``HighThroughputEngine``.  This class was deprecated in :ref:`v2.27.0
+  <changelog-2.27.0>`.  We recommend migrating relevant configurations to use
+  |GlobusComputeEngine|.
