+++ compute_endpoint/globus_compute_endpoint/engines/__init__.py
@@ -1,6 +1,5 @@
 from .globus_compute import GlobusComputeEngine
 from .globus_mpi import GlobusMPIEngine
-from .high_throughput.engine import HighThroughputEngine
 from .process_pool import ProcessPoolEngine
 from .thread_pool import ThreadPoolEngine
 
@@ -9,5 +8,4 @@ __all__ = (
     "GlobusMPIEngine",
     "ProcessPoolEngine",
     "ThreadPoolEngine",
-    "HighThroughputEngine",
 )
