--- compute_endpoint/globus_compute_endpoint/engines/high_throughput/mac_safe_queue.py
@@ -1,12 +1,0 @@
-import multiprocessing as mp
-import platform
-import typing as t
-
-mpQueue: t.Type[mp.Queue]
-
-if platform.system() == "Darwin":
-    from parsl.multiprocessing import MacSafeQueue as mpQueue
-else:
-    from multiprocessing import Queue as mpQueue
-
-__all__ = ("mpQueue",)
